Event targetting now caches event target for reposition method
PRior to this commit when using 'event' for the position target, calling the reposition
method with no event object would cause an error, since the method had no target reference
to use for positioning calculations. Previous target is now used if no event object is provided
to prevent this issue.