Removed inconsistency in hoverIntent code that added mouseleave event to hide events
Originally the code added a mouseleave event that cleared the show timer
when the show target was first moused over. This was inconsistent with futher
show events where the mouseleave would NOT clear the timer, leading to confusion
over the show.delay behaviour. If you are using a special event like 'unfocus'
that requires a hoverIntent type behaviour, you must added it yourself via some
custom code.