Ensure all qtips are destroyed when elements are removed in any manor (.html/.empty/.remove)
Remove $.fn.remove override for much better $.cleanData override as the jQuery UI
plugin does. This makes sure that all elements with a qTip bound that are destroyed
by a jQuery call to .html, .empty, .remove etc always get removed. Also ensures that
this override only takes place if jQuery UI isn't present, as it does this for us.