Fix redraw() issues when body has complex flows
Instead of just applying a fluidClass, append the tooltips to a container
outside the normal document flow, and float them as before using the old
fluidClass. This prevents problems with width/height calculation when the
document has complex flows with min/max-width and inline-block elements.

Also remove log() function and it's calls, as they aren't really needed
and just cause console flooding in a lot of cases.

Tips plugin has a small refactor to fix it in regards to the above
redraw() changes.