Fix IE tip not showing bug and cleanup tip/core.js source.
Added new tooltipredraw(n) events, which are used to fix the IE
tip invisibility problem when using redraw(). Cleaned up core source
with common triggerEvent method to reduce code duplication and added
similar createVML method to tips.js to abstract that out.

Also changed the default vml namespace to qvml (from vml) and add default vml class
of .qtip-vml, which is now styled in tip.css to give it the default
VMML behavior. This prevents IE from not showing it if the tips are
for some reason removed/re-appended to the DOM.