Fix tooltip position calculation for SVG circles
The tooltip position for SVG circles is not correct if the SVG has a viewBox set. The radius is not adjusted to the zoom the viewBox results to.

In the polys the position for "center" is also wrong, it should return the center of the circle but the x coordinate is wrong.

This fixes both.