set('content.text', val) doesn't change content.
When I tried updating the content through the API it didn't actually update the content. Every time the option got removed because the tooltip was rendered and the option not either show or prerender.

I think you probably meant to remove the options 'prerender' or 'show' if the qtip was already rendered.